requirements.txt
streamlit
langchain-community
langchain-google-genai
langchain-core
faiss-cpu
requests
python-dotenv
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-core